==> 2_Get_WikiData.ipynb <==
# Getting Data from Wikidata for each economist

The code below loads the economists list, query Wikidata for each QID, extracts gender, birth year, country of citizenship, occupations, field of work and saves results into a new CSV. The CSV will be used later to merge with pageview data in Step 3

import pandas as pd
import requests
import time

df = pd.read_csv("economists_list_with_summaries.csv")
df.head()

def fetch_wikidata(qid):
    """Fetch gender, birth year, citizenship, occupations, field of work from Wikidata."""
    if pd.isna(qid):
        return None

    url = f"https://www.wikidata.org/wiki/Special:EntityData/{qid}.json"
    headers = {"User-Agent": "SohaKhan/1.0 (sk131@wellesley.edu) CourseProject/CS234; respectful academic use"}

    try:
        r = requests.get(url, headers=headers, timeout=10)
        data = r.json()
    except Exception as e:
        print("Error:", qid, e)
        return None

    try:
        ent = data["entities"][qid]["claims"]
    except KeyError:
        return None
    
    def get_value(prop):
        """Extract value from wikidata claim."""
        if prop not in ent:
            return None
        values = ent[prop]
        results = []
        for v in values:
            datav = v["mainsnak"].get("datavalue", {})
            if "value" in datav:
                results.append(datav["value"])
        return results if results else None
    
    # gender (P21)
    gender = get_value("P21")
    # birthdate (P569)
    birthdate = get_value("P569")
    # citizenship (P27)
    citizenship = get_value("P27")
    # occupations (P106)
    occupations = get_value("P106")
    # field of work (P101)
    fields = get_value("P101")

    return {
        "gender_raw": gender,
        "birthdate_raw": birthdate,
        "citizenship_raw": citizenship,
        "occupation_raw": occupations,
        "field_raw": fields
    }


wikidata_results = []

for i, row in df.iterrows():
    qid = row["qid"]
    if pd.isna(qid):
        wikidata_results.append({})
        continue

    info = fetch_wikidata(qid)
    wikidata_results.append(info)
    
    if i % 50 == 0:
        print(f"Processed {i} economists...")
    
    time.sleep(0.1)


wikidata_df = pd.DataFrame(wikidata_results)
full = pd.concat([df, wikidata_df], axis=1)
full.head()

# Birth year
def extract_year(birth_raw):
    if not birth_raw or not isinstance(birth_raw, list):
        return None
    value = birth_raw[0].get("time")   # format: "+1971-00-00T00:00:00Z"
    if value:
        return int(value[1:5])
    return None

full["birth_year"] = full["birthdate_raw"].apply(extract_year)

# Gender
def extract_gender(gender_raw):
    if not gender_raw:
        return None
    try:
        return gender_raw[0]["id"]   # Q6581097 = male, Q6581072 = female
    except:
        return None

full["gender_qid"] = full["gender_raw"].apply(extract_gender)


full.to_csv("economists_with_wikidata.csv", index=False)
