==> 4_Get_2024_Pageviews.ipynb <==
import duckdb
import pandas as pd

# Load the cleaned economist metadata
econ = pd.read_csv("economists_master_dataset.csv")

# Extract article title from URL
econ["page_title"] = econ["article_url_x"].str.split("/wiki/").str[-1]

econ.head()


econ = pd.read_csv("economists_master_dataset.csv")

conn = duckdb.connect()
conn

conn.execute("INSTALL httpfs;")
conn.execute("LOAD httpfs;")

database_url = "https://cs.wellesley.edu/~eni/duckdb/2024_wiki_views.duckdb"

try:
    conn.execute(f"ATTACH '{database_url}' AS web_db (READ_ONLY);")
    conn.execute("USE web_db;")
    print(f"Successfully attached database from: {database_url}")
except Exception as e:
    print(f"Error attaching database: {e}")

query = "PRAGMA show_tables"
result = conn.sql(query)
result

table_name = "data_table"
query = f"PRAGMA table_info('web_db.{table_name}');"

# We can apply the method .df() to the result of the query to convert it into a dataframe
column_info_df = conn.sql(query).df()
column_info_df

query_1 = """
SELECT * FROM data_table
LIMIT 10;
"""
result_1 = conn.sql(query_1).df() # after executing, convert to df for better printout

result_1

econ_qids = duckdb.from_df(econ[["qid"]])

conn.register("econ_df", econ)

conn.execute("""
    CREATE TEMP TABLE econ_q AS 
    SELECT qid FROM econ_df
""")

query = """
SELECT 
    e.qid,
    SUM(d.pageviews) AS total_views_2024,
    d.date
FROM econ_q e
LEFT JOIN data_table d
    ON e.qid = d.qid
GROUP BY e.qid
ORDER BY total_views_2024 DESC;
"""
views_2024 = conn.execute(query).df()
views_2024.head()


econ_with_views = econ.merge(views_2024, on="qid", how="left")
econ_with_views.head()

econ_with_views.to_csv('merged_econ_data.csv', index=False)
